requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.8
django-crispy-forms==1.6.1
djangorestframework==3.6.3
gunicorn==19.7.1
idna==2.5
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3
requests==2.18.2
urllib3==1.22
whitenoise==3.3.0
